+++ common/src/main/java/net/mehvahdjukaar/amendments/common/block/BoilingWaterCauldronBlock.java
@@ -88,7 +88,7 @@ public class BoilingWaterCauldronBlock extends LayeredCauldronBlock {
     public BlockState updateShape(BlockState state, Direction direction, BlockState neighborState, LevelAccessor level, BlockPos currentPos, BlockPos neighborPos) {
         var s = super.updateShape(state, direction, neighborState, level, currentPos, neighborPos);
         if (direction == Direction.DOWN) {
-            boolean isFire = LiquidCauldronBlock.shouldBoil(neighborState, SoftFluidStack.of(BuiltInSoftFluids.WATER.getHolder()),
+            boolean isFire = LiquidCauldronBlock.shouldBoil(neighborState, SoftFluidStack.of(BuiltInSoftFluids.WATER),
                     level, neighborPos);
             s = s.setValue(BOILING, isFire);
         }