+++ common/src/main/java/net/mehvahdjukaar/amendments/common/block/FloorCandleSkullBlock.java
@@ -89,4 +89,6 @@ public class FloorCandleSkullBlock extends AbstractCandleSkullBlock implements I
         }
         return false;
     }
+
+
 }