+++ common/src/main/java/net/mehvahdjukaar/amendments/reg/ModRegistry.java
@@ -40,7 +40,6 @@ import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.RecipeSerializer;
-import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.block.BannerBlock;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
@@ -102,7 +101,7 @@ public class ModRegistry {
             res("dye_bottle"), DyeBottleRecipe::new);
 
     public static final Supplier<MenuType<LecternEditMenu>> LECTERN_EDIT_MENU = RegHelper.registerMenuType(
-            res("lectern_edit"), LecternEditMenu::new
+            res("lectern_edit"), LecternEditMenu::of
     );
 
     public static final RegSupplier<SimpleParticleType> BOILING_PARTICLE = RegHelper.registerParticle(res("boiling_bubble"));