+++ changelog.md
@@ -1,2 +1,2 @@
-added more tags for many modded items (thanks to Cicopath)
-added skull pile blacklist tag
+- fixed lectern editing not working properly
+- updated to newest moonlight version